email verification banner

+++ pages/api/activities/delete.ts
@@ -1,6 +1,12 @@
 import { validateRoute, prisma } from 'lib';
 
 export default validateRoute(async (req, res, user) => {
+  // Don't allow if user is not verified
+  if (!user.isVerified) {
+    res.status(403);
+    res.json({ error: 'Email not verified' });
+    return;
+  }
   try {
     const deletedActivity = await prisma.activity.delete({
       where: { id: req.body.id },
